refactor: 패키지 위치 변경 및 클래스 네이밍 변경
