+++ tomcat/src/main/java/org/apache/coyote/http11/handler/adapter/HandlerAdapter.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.handler.adapter;
 
-import nextstep.jwp.controller.base.Controller;
+import org.apache.coyote.http11.handler.mapper.controller.Controller;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
