+++ tomcat/src/main/java/org/apache/coyote/http11/request/header/Headers.java
@@ -8,7 +8,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import static org.apache.coyote.http11.Constant.HEADER_SEPARATOR;
+import static org.apache.coyote.http11.common.Constant.HEADER_SEPARATOR;
 
 public class Headers {
