+++ tomcat/src/test/java/org/apache/coyote/http11/handler/controller/base/RequestPathTest.java
@@ -1,7 +1,7 @@
 package org.apache.coyote.http11.handler.controller.base;
 
-import nextstep.jwp.controller.base.Dict;
 import nextstep.jwp.exception.NotFoundException;
+import org.apache.coyote.http11.handler.mapper.controller.RequestPath;
 import org.apache.coyote.http11.request.uri.HttpMethod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
-class DictTest {
+class RequestPathTest {
 
     @DisplayName("path와 HttpMethod를 이용해서 Enum을 가져온다")
     @Test
@@ -19,10 +19,10 @@ class DictTest {
         HttpMethod httpMethod = HttpMethod.GET;
 
         // when
-        Dict dict = Dict.find(path, httpMethod);
+        RequestPath requestPath = RequestPath.find(path, httpMethod);
 
         // then
-        assertThat(dict).isEqualTo(Dict.INDEX_GET);
+        assertThat(requestPath).isEqualTo(RequestPath.INDEX_GET);
     }
 
     @DisplayName("path혹은 HttpMethod가 잘못된다면 예외를 발생한다")
@@ -38,10 +38,10 @@ class DictTest {
 
         // when & then
         assertSoftly(softly -> {
-            softly.assertThatThrownBy(() -> Dict.find(path, invalidHttpMethod))
+            softly.assertThatThrownBy(() -> RequestPath.find(path, invalidHttpMethod))
                     .isInstanceOf(NotFoundException.class);
 
-            softly.assertThatThrownBy(() -> Dict.find(invalidPath, httpMethod))
+            softly.assertThatThrownBy(() -> RequestPath.find(invalidPath, httpMethod))
                     .isInstanceOf(NotFoundException.class);
         });
     }
